Hotfix for HWID bug
A message can accidentally reset the HWID or attempt to clear the HWID fuse repeatedly, hanging a radio. Why this occurs repeatedly is unclear and may result from burning out an EEPROM location, as it's unclear how long this bug has existed and how many EEPROM writes were attempted on the radio it was observed on. This commit attempts to isolate HWID behavior such that it cannot be triggered by packets received over radio. Consider deprecating HWID in a future release